Refactor into subpanels
