+++ ui/space_types.py
@@ -1,0 +1,1 @@
+SPACE_TYPES = {'IMAGE_EDITOR', 'NODE_EDITOR'}
